requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
mypy==1.16.0
mypy_extensions==1.1.0
pathspec==0.12.1
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
requests==2.32.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
watchdog==6.0.0
